[feat] Network 컴포넌트 분리 및 링크 공유, 홈, 폴더 디테일 뷰에서 Network 컴포넌트 적용

+++ app/src/main/java/com/hyeeyoung/wishboard/presentation/home/WishListViewModel.kt
@@ -2,15 +2,18 @@ package com.hyeeyoung.wishboard.presentation.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.hyeeyoung.wishboard.data.model.folder.FolderItem
 import com.hyeeyoung.wishboard.data.model.wish.WishItem
 import com.hyeeyoung.wishboard.domain.repositories.CartRepository
 import com.hyeeyoung.wishboard.domain.repositories.FolderRepository
 import com.hyeeyoung.wishboard.domain.repositories.WishRepository
+import com.hyeeyoung.wishboard.presentation.base.viewmodel.NetworkViewModel
 import com.hyeeyoung.wishboard.presentation.cart.types.CartStateType
+import com.hyeeyoung.wishboard.util.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -19,15 +22,18 @@ class WishListViewModel @Inject constructor(
     private val wishRepository: WishRepository,
     private val cartRepository: CartRepository,
     private val folderRepository: FolderRepository,
-) : ViewModel() {
+) : NetworkViewModel() {
+    private val _wishListFetchState = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
+    val wishListFetchState get() = _wishListFetchState.asStateFlow()
     private val wishList = MutableLiveData<List<WishItem>?>(listOf())
     private val wishListAdapter = WishListAdapter()
-    private var folderItem: FolderItem? = null
+    private var _folderItem: FolderItem? = null
+    val folderItem get() = _folderItem
 
     fun fetchWishList() {
         viewModelScope.launch {
             wishRepository.fetchWishList().let {
-                wishList.postValue(it)
+                wishList.value = it
                 wishListAdapter.setData(it)
             }
         }
@@ -43,22 +49,25 @@ class WishListViewModel @Inject constructor(
         if (folderId == null) return
         viewModelScope.launch {
             folderRepository.fetchItemsInFolder(folderId).let { folders ->
-                wishList.postValue(folders)
+                _wishListFetchState.value =
+                    if (folders == null) UiState.Error(null) else UiState.Success(true)
+                wishList.value = folders
                 wishListAdapter.setData(folders)
             }
         }
     }
 
     fun toggleCartState(position: Int, item: WishItem) {
+        if (!isConnected.value) return
         viewModelScope.launch {
             val isInCart = item.cartState == CartStateType.IN_CART.numValue
             val isSuccessful =
                 if (isInCart) {
-                    cartRepository.removeToCart( item.id ?: return@launch)
+                    cartRepository.removeToCart(item.id ?: return@launch)
                 } else {
                     cartRepository.addToCart(item.id ?: return@launch)
                 }
-            if (!isSuccessful) return@launch // TODO 네트워크 연결 실패, 그 외 나머지 예외 처리 -> 스낵바 띄우기
+            if (!isSuccessful) return@launch
 
             item.also { wishItem ->
                 wishItem.cartState = toggleCartState(wishItem.cartState)
@@ -90,10 +99,9 @@ class WishListViewModel @Inject constructor(
     }
 
     fun setFolderItem(folderItem: FolderItem) {
-        this.folderItem = folderItem
+        _folderItem = folderItem
     }
 
     fun getWishList(): LiveData<List<WishItem>?> = wishList
     fun getWishListAdapter(): WishListAdapter = wishListAdapter
-    fun getFolderItem(): FolderItem? = folderItem
 }
